Remove icon from dropdown menu
