+++ src/pages/Exchange/MainArea/Orders/index.tsx
@@ -35,7 +35,6 @@ const Orders: SFC = ({className}) => {
       const menuOptions = [
         {
           label: 'View Trades',
-          menuIcon: '',
           onClick: () => {
             setSelectedOrder(order);
             toggleTradesModal();
@@ -46,7 +45,6 @@ const Orders: SFC = ({className}) => {
       if ([FillStatus.OPEN, FillStatus.PARTIALLY_FILLED].includes(order.fill_status)) {
         menuOptions.unshift({
           label: 'Cancel Order',
-          menuIcon: '',
           onClick: async () => {
             await dispatch(updateExchangeOrder(order.id, {fill_status: FillStatus.CANCELLED}));
           },
